A final update on what i learnt today
I learnt about Pipelines and its benefits when building a model.

I built a pipeline for preprocessing numerical features in the dataset.

Then i built a pipeline for preprocessing categorical features in the dataset.

I then learnt about a transformer that's like a super pipeline that can preprocess both categorical and numerical features in a dataset.