a little update
So i discovered that the ColumnTransformer didn't effectively clean the missing the data in the housing dataset.

I decided to build a custom transformer and it was able to effectively clean the dataset which resulted to no missing values in the dataset.

I then tested this custom ColumnTransformer  on the titanic dataset, and it effectively cleaned the dataset.

I'm still going to learn more about Machine Learning model building and deployment.