Sat Jan 27 Lesson Update
I trained LinearRegression on the data (using the preprocessing pipeline and the housing_labels) and i calculated the LinearRegression model RMSE using sklearn mean_squared_errror.  The Root Mean Squared Error is 68647.95 and this is high and could be due to underfitting.

The next model that i used is the DecisionTreeRegressor. This model RMSE was 0.0 and this can't be relied on, there might be chances of overfitting.

The next thing that i applied was Sklearn Cross Validation Score. When calculating cross_val_score, apply - in front of it ( -cross_val_score(tree_reg, housing, housing_labels, scoring="neg_root_mean_squared_error", cv=10)

The decision tree has an RMSE of about 67146, with a standard
deviation of about 2125

There’s an overfitting problem because the training error is low (actually zero)
while the validation error is high

The next model that was trained is the RandomForestRegressor.
Its RMSE was these:
- From CrossValidation: 47016
- From mean_squared_error: 17547

The next thing that i applied was fine tuning of machine learning models.

I learnt and applied GridSearchCV.

The process involved in GridSearchCV are these: Define Hyperparameters, Define Parameter Grid, Cross-validation, Select Best Hyperparameters.

After building the GridSearchCV, i then fit it to the housing and housing_labels.

To get the best parameters from the GridSearchCV, I used this: best_params_

The best parameters were these: {'preprocessing__geo__n_clusters': 15, 'random_forest__max_features': 6}

That's all for today from Hands On Machine Learning.

I need to continue with my DataCamp Professional Data Scientist Track Today.

I aim to be a very good Data Scientist and ML Scientist/Engineer